Move linux account manager to accounts.go
